Elevator Trigger code that might work maybe

+++ 2018 PowerUp/src/org/usfirst/frc/team4043/robot/subsystems/Elevator.java
@@ -12,7 +12,7 @@ public class Elevator extends Subsystem {
     // Put methods for controlling this subsystem
     // here. Call these from Commands.
 	
-	public void elevatorMove(float axis) {
+	public void elevatorMove(double axis) {
 		Robot.elevatorPID.disable();
 		RobotMap.evelator.set(axis);
 	}
